refactor: post is now a component
